Feat: 시간 타이머 연동 및 PutPiece 로직 연동

+++ Assets/Scripts/data/GameData.cs
@@ -12,7 +12,7 @@ public class GameData
     /*
      * "turnedAt":[2024,7,14,18,45,7,868696000]
      */
-    public long[] turnedAt;
+    public int[] turnedAt;
     public string winnerPlayerId;
     public Piece[] board;
 }
